iOS: Fix test action picking the wrong scheme (#3)
Xcode picks the wrong scheme because there are two ReactTestApp schemes:
One is the test app itself (which we want), and the other is the Pod
that only contains public interfaces. The latter will eventually be
superseded by using native modules but for now, we will rename it to
avoid confusing Xcode.