fix(Android): Wrong version of Hermes is bundled in APK
The wrong version of Hermes can be bundled in the APK if multiple
versions are installed and the wrong version is being hoisted. For
instance, bumping `react-native` to ^0.63 and keeping
`react-native-macos` at ^0.62, will install both `hermes-engine@0.4.3`
and `hermes-engine@0.5.1`, where the former is hoisted. We fix this by
using `react-native`'s `node_modules` folder as starting point when
looking for `hermes-engine`.