fix(macos): `react-native-macos` canary breaks `pod install` (#421)
A recent change in `react-native-macos` is assuming that an Xcode
project file lives next to the `Podfile`, causing `flipper_post_install`
to fail. Since the script currently doesn't do anything that we aren't
already handling, and it doesn't look like the change exists upstream,
we'll ignore the error for now.