fix: make @react-native-community/cli optional (#632)
Package managers issue a warning when `@react-native-community/cli`
isn't explicitly provided by consumers. We're making it optional because
we can assume that it gets installed with react-native. If this changes,
we will have to make changes anyway.

This will suppress the warnings, and it will also avoid having to keep
`@react-native-community/cli` aligned with the installed version of
`react-native`.