test: introduce e2e tests (#1467)
There are a number of issues we need to address before we can add these
to CI. For instance, Android emulators can spend 5 minutes just to boot
up, and they require either a Windows or macOS agent. For now, we will
only use e2e tests to automate some of the manual testing against newer
versions of React Native.